+++ Sources/Extensions/UIKit/UITextFieldExtension.swift
@@ -31,7 +31,7 @@ extension UITextField {
 
 // MARK: - Placeholder
 
-public extension UITextField {
+extension UITextField {
 
     public func setPlaceHolderTextColor(_ color: UIColor) {
         guard let placeholder = placeholder, placeholder.isNotEmpty else {